ext4 crypto: add support for hardware accelerated encryption
This is an initial check in to support hw accelerated encryption using
Qualcomm's ICE hardware.

This is not intended for upstream merging, since it depends on
interfaces in Qualcomm's BSP kernel that are unlikely to be accepted
upstream without significant cleanup.

Signed-off-by: Michael Halcrow <mhalcrow@google.com>
Signed-off-by: "Theodore Ts'o" <tytso@google.com>
Change-Id: I8cd5fe72e50e460b2d44cbd6cdaa443dcacc2288

ext4: ICE key type and reference kernel keyring

Introduces ICE key type/mode (127). With this change, both the policy in
the xattr and the key in the keyring must be type/mode 127 to use QC
ICE. Else, ext4 crypto will use software encryption.

Changes the key used in QC PFK from a master key copied into each struct
bio to a reference to a key in the keyring. The key should be present so
long as its ext4_crypt_info container exists, which should be for the
life of the block I/O.

Change-Id: I445936255e74612d33272c8d490d0a88cf9e3295

ext4: Don't use ICE for symlink pages

Change-Id: I2b6935665f9e9a2badc4cf76f694c5f876869204

ext4: Hack ICE fix to set the bio encryption context when performing a
low-level read when we need to pull the page into the cache on the write
path.

bug 28995264

Change-Id: I97fbb22299546615919990bf1e60b3880214a031